+++ Scripts/CombatEntityStats.cs
@@ -1,7 +1,7 @@
 ﻿public record CombatEntityStats
 {
-    public float MaxHealth { get; set; }
-    public float CurrentHealth { get; set; }
+    public double MaxHealth { get; set; }
+    public double CurrentHealth { get; set; }
     public int AttackDamage { get; set; }
     public int Speed { get; set; }
 
@@ -23,7 +23,7 @@
             Speed = a.Speed + b.Speed
         };
 
-    public static CombatEntityStats operator *(CombatEntityStats stats, float mul) =>
+    public static CombatEntityStats operator *(CombatEntityStats stats, double mul) =>
         new CombatEntityStats()
         {
             MaxHealth = stats.MaxHealth * mul,
